backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.30
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pydantic[email-validator]==2.7.4
pydantic-settings==2.3.4
redis==5.0.7
celery==5.4.0
httpx==0.27.0
cryptography==42.0.8
MetaTrader5==5.0.4230
